requirements.txt
aistudio-sdk==0.3.8
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bce-python-sdk==0.9.56
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorclass==2.2.2
colorlog==6.10.1
compressed-rtf==1.0.7
cryptography==46.0.3
deprecation==2.1.0
easygui==0.98.3
easyocr==1.7.2
ebcdic==1.1.1
et_xmlfile==2.0.0
extract-msg==0.55.0
filelock==3.20.1
fsspec==2025.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lance-namespace==0.4.0
lance-namespace-urllib3-client==0.4.0
lancedb==0.26.0
lark==1.3.1
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
modelscope==1.33.0
mpmath==1.3.0
msoffcrypto-tool==5.4.2
narwhals==2.14.0
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
olefile==0.47
oletools==0.60.2
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.12
pandas==2.3.3
pcodedmp==1.2.6
pillow==12.0.0
prettytable==3.17.0
protobuf==6.33.2
psutil==7.2.0
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyMuPDF==1.26.7
pyparsing==3.3.1
pypdfium2==5.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-magic==0.4.27
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
red-black-tree-mod==1.22
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
RTFDE==0.1.2.2
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.12.20
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.2
watchdog==6.0.0
wcwidth==0.2.14
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
